1. GOAP 8.10.2021: GOAP system implemented
But it ain't very good, or rather it isn't smart system. It's just two while loops and if statements. With this system, the user would have to play lot with  the values.  Also, this might be too bloated and expanding the system is gonna be hard.

Really need to find some better inspiration for this.